requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==25.2.0
inflection==0.5.1
np==1.0.2
numpy==1.26.4
packaging==24.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
